backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "lancedb>=0.22.0",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.2.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "pyvips>=3.0.0",
    "safetensors>=0.5.3",
    "tokenizers>=0.21.1",
    "torch>=2.7.0",
]
